STACK CACHE
Added cache property to stack's push and pop methods.
If you cache a transitioned view the element is kept in the DOM, when the view
is recalled to top of the stack it's render() function is not invoked.

The result is that the state of the DOM is kept intact at time of transition. But,
if the element has animation, video, or sound they'll continue to run.